refactor: split native agent runtime handlers
